Group triggers by service
