+++ functions/src/firestore/emitCreatedContactEvent.ts
