+++ functions/src/firestore/notifyInfectedUser.ts
